refactor!: restructure into clean architecture
- Move internals to business/model and service layers
- Remove CJS dual-output and lib/ directory
- Replace stringUtil with httpResponseCodeUtil
- Add contract tests for models and utils

BREAKING CHANGE: Sub-path imports have changed.
See MIGRATION.md for full details.
