+++ .codegraph/daemon.pid
@@ -1,0 +1,6 @@
+{
+  "pid": 811967,
+  "version": "0.9.9",
+  "socketPath": "/home/milos/code/msh/packages/error/.codegraph/daemon.sock",
+  "startedAt": 1780952358415
+}
