+++ .dockerignore
@@ -9,3 +9,4 @@
 /test/node_modules
 /test/dist
 *.tsbuildinfo
+.history
