+++ .husky/commit-msg
@@ -1,1 +1,4 @@
+export NVM_DIR="${NVM_DIR:-$HOME/.nvm}"
+[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
+
 npx --no-install commitlint --edit "$1"
