--- .lintstagedrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-	'**/*.json': ['npm run lint-json'],
-	'src/**/*.{ts,tsx,js,jsx}': ['eslint', 'prettier --check'],
-}
