+++ .prettierignore
@@ -7,3 +7,4 @@
 /resource
 /packages
 package-lock.json
+.history
