+++ eslint.config.js
@@ -4,22 +4,22 @@ export default [
 	{
 		ignores: [
 			'.base-frame-template',
+			'.history',
 			'.idea',
 			'.semaphore',
 			'coverage',
 			'dist',
-			'lib',
 			'node_modules',
 			'resource',
 			'packages',
 			'.*.js',
+			'.*.cjs',
 			'.*.json',
 			'package-lock.json',
 			'tsconfig*.json',
 			'typedoc.json',
 			'jest.config*.ts',
 			'eslint.config.js',
-			'babel.config.cjs',
 			'release.config.cjs',
 		],
 	},
