+++ resource/doc/api/README.md
@@ -1,16 +1,16 @@
-@beecode/msh-error
+**@beecode/msh-error**
+
+***
 
 # @beecode/msh-error
 
-## Table of contents
+## Modules
 
-### Modules
-
-- [error/base-model](modules/error_base_model.md)
-- [error/factory](modules/error_factory.md)
-- [error/hof](modules/error_hof.md)
-- [express/middleware](modules/express_middleware.md)
-- [http-response-code](modules/http_response_code.md)
-- [index](modules/index.md)
-- [types/global](modules/types_global.md)
-- [util/string-util](modules/util_string_util.md)
+- [business/model/error-base-model](business/model/error-base-model/README.md)
+- [business/model/http-response-code-mapper](business/model/http-response-code-mapper/README.md)
+- [business/service/error-factory](business/service/error-factory/README.md)
+- [business/service/error-hof](business/service/error-hof/README.md)
+- [business/service/express-middleware](business/service/express-middleware/README.md)
+- [index](index/README.md)
+- [types/global](types/global/README.md)
+- [util/http-response-code-util](util/http-response-code-util/README.md)
