+++ resource/doc/api/business/model/error-base-model/README.md
@@ -1,0 +1,15 @@
+[**@beecode/msh-error**](../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../README.md) / business/model/error-base-model
+
+# business/model/error-base-model
+
+## Classes
+
+- [ErrorBaseModel](classes/ErrorBaseModel.md)
+
+## Type Aliases
+
+- [ErrorPayload](type-aliases/ErrorPayload.md)
