+++ resource/doc/api/business/model/error-base-model/classes/ErrorBaseModel.md
@@ -1,0 +1,265 @@
+[**@beecode/msh-error**](../../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../../README.md) / [business/model/error-base-model](../README.md) / ErrorBaseModel
+
+# Class: ErrorBaseModel\<T\>
+
+Defined in: [packages/error/src/business/model/error-base-model.ts:7](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/business/model/error-base-model.ts#L7)
+
+## Extends
+
+- `Error`
+
+## Type Parameters
+
+### T
+
+`T` *extends* [`ErrorPayload`](../type-aliases/ErrorPayload.md) = `any`
+
+## Constructors
+
+### Constructor
+
+> **new ErrorBaseModel**\<`T`\>(`params`): `ErrorBaseModel`\<`T`\>
+
+Defined in: [packages/error/src/business/model/error-base-model.ts:11](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/business/model/error-base-model.ts#L11)
+
+#### Parameters
+
+##### params
+
+###### httpCode
+
+[`HttpResponseCodeMapper`](../../http-response-code-mapper/enumerations/HttpResponseCodeMapper.md)
+
+###### message?
+
+`string`
+
+###### name?
+
+`string`
+
+###### payload?
+
+`T`
+
+#### Returns
+
+`ErrorBaseModel`\<`T`\>
+
+#### Overrides
+
+`Error.constructor`
+
+## Properties
+
+### cause?
+
+> `optional` **cause?**: `unknown`
+
+Defined in: node\_modules/typescript/lib/lib.es2022.error.d.ts:24
+
+#### Inherited from
+
+`Error.cause`
+
+***
+
+### httpCode
+
+> **httpCode**: `number`
+
+Defined in: [packages/error/src/business/model/error-base-model.ts:8](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/business/model/error-base-model.ts#L8)
+
+***
+
+### message
+
+> **message**: `string`
+
+Defined in: node\_modules/typescript/lib/lib.es5.d.ts:1075
+
+#### Inherited from
+
+`Error.message`
+
+***
+
+### name
+
+> **name**: `string`
+
+Defined in: node\_modules/typescript/lib/lib.es5.d.ts:1074
+
+#### Inherited from
+
+`Error.name`
+
+***
+
+### payload?
+
+> `optional` **payload?**: `T`
+
+Defined in: [packages/error/src/business/model/error-base-model.ts:9](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/business/model/error-base-model.ts#L9)
+
+***
+
+### stack?
+
+> `optional` **stack?**: `string`
+
+Defined in: node\_modules/typescript/lib/lib.es5.d.ts:1076
+
+#### Inherited from
+
+`Error.stack`
+
+***
+
+### stackTraceLimit
+
+> `static` **stackTraceLimit**: `number`
+
+Defined in: node\_modules/@types/node/globals.d.ts:67
+
+The `Error.stackTraceLimit` property specifies the number of stack frames
+collected by a stack trace (whether generated by `new Error().stack` or
+`Error.captureStackTrace(obj)`).
+
+The default value is `10` but may be set to any valid JavaScript number. Changes
+will affect any stack trace captured _after_ the value has been changed.
+
+If set to a non-number value, or set to a negative number, stack traces will
+not capture any frames.
+
+#### Inherited from
+
+`Error.stackTraceLimit`
+
+## Methods
+
+### captureStackTrace()
+
+> `static` **captureStackTrace**(`targetObject`, `constructorOpt?`): `void`
+
+Defined in: node\_modules/@types/node/globals.d.ts:51
+
+Creates a `.stack` property on `targetObject`, which when accessed returns
+a string representing the location in the code at which
+`Error.captureStackTrace()` was called.
+
+```js
+const myObject = {};
+Error.captureStackTrace(myObject);
+myObject.stack;  // Similar to `new Error().stack`
+```
+
+The first line of the trace will be prefixed with
+`${myObject.name}: ${myObject.message}`.
+
+The optional `constructorOpt` argument accepts a function. If given, all frames
+above `constructorOpt`, including `constructorOpt`, will be omitted from the
+generated stack trace.
+
+The `constructorOpt` argument is useful for hiding implementation
+details of error generation from the user. For instance:
+
+```js
+function a() {
+  b();
+}
+
+function b() {
+  c();
+}
+
+function c() {
+  // Create an error without stack trace to avoid calculating the stack trace twice.
+  const { stackTraceLimit } = Error;
+  Error.stackTraceLimit = 0;
+  const error = new Error();
+  Error.stackTraceLimit = stackTraceLimit;
+
+  // Capture the stack trace above function b
+  Error.captureStackTrace(error, b); // Neither function c, nor b is included in the stack trace
+  throw error;
+}
+
+a();
+```
+
+#### Parameters
+
+##### targetObject
+
+`object`
+
+##### constructorOpt?
+
+`Function`
+
+#### Returns
+
+`void`
+
+#### Inherited from
+
+`Error.captureStackTrace`
+
+***
+
+### isError()
+
+> `static` **isError**(`error`): `error is Error`
+
+Defined in: node\_modules/typescript/lib/lib.esnext.error.d.ts:21
+
+Indicates whether the argument provided is a built-in Error instance or not.
+
+#### Parameters
+
+##### error
+
+`unknown`
+
+#### Returns
+
+`error is Error`
+
+#### Inherited from
+
+`Error.isError`
+
+***
+
+### prepareStackTrace()
+
+> `static` **prepareStackTrace**(`err`, `stackTraces`): `any`
+
+Defined in: node\_modules/@types/node/globals.d.ts:55
+
+#### Parameters
+
+##### err
+
+`Error`
+
+##### stackTraces
+
+`CallSite`[]
+
+#### Returns
+
+`any`
+
+#### See
+
+https://v8.dev/docs/stack-trace-api#customizing-stack-traces
+
+#### Inherited from
+
+`Error.prepareStackTrace`
