+++ resource/doc/api/business/model/error-base-model/type-aliases/ErrorPayload.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-error**](../../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../../README.md) / [business/model/error-base-model](../README.md) / ErrorPayload
+
+# Type Alias: ErrorPayload
+
+> **ErrorPayload** = `Record`\<`string`, `unknown`\>
+
+Defined in: [packages/error/src/business/model/error-base-model.ts:4](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/business/model/error-base-model.ts#L4)
