+++ resource/doc/api/business/model/http-response-code-mapper/README.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-error**](../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../README.md) / business/model/http-response-code-mapper
+
+# business/model/http-response-code-mapper
+
+## Enumerations
+
+- [HttpResponseCodeMapper](enumerations/HttpResponseCodeMapper.md)
