+++ resource/doc/api/business/service/error-factory/README.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-error**](../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../README.md) / business/service/error-factory
+
+# business/service/error-factory
+
+## Functions
+
+- [errorFactory](functions/errorFactory.md)
