+++ resource/doc/api/business/service/error-hof/README.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-error**](../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../README.md) / business/service/error-hof
+
+# business/service/error-hof
+
+## Functions
+
+- [errorHOF](functions/errorHOF.md)
