+++ resource/doc/api/business/service/error-hof/functions/errorHOF.md
@@ -1,0 +1,21 @@
+[**@beecode/msh-error**](../../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../../README.md) / [business/service/error-hof](../README.md) / errorHOF
+
+# Function: errorHOF()
+
+> **errorHOF**(`httpCode`): \{\<`T`\>(`message?`, `payload?`): [`ErrorBaseModel`](../../../model/error-base-model/classes/ErrorBaseModel.md); \<`T`\>(`params?`): [`ErrorBaseModel`](../../../model/error-base-model/classes/ErrorBaseModel.md); \}
+
+Defined in: [packages/error/src/business/service/error-hof.ts:5](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/business/service/error-hof.ts#L5)
+
+## Parameters
+
+### httpCode
+
+[`HttpResponseCodeMapper`](../../../model/http-response-code-mapper/enumerations/HttpResponseCodeMapper.md)
+
+## Returns
+
+\{\<`T`\>(`message?`, `payload?`): [`ErrorBaseModel`](../../../model/error-base-model/classes/ErrorBaseModel.md); \<`T`\>(`params?`): [`ErrorBaseModel`](../../../model/error-base-model/classes/ErrorBaseModel.md); \}
