+++ resource/doc/api/business/service/express-middleware/README.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-error**](../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../README.md) / business/service/express-middleware
+
+# business/service/express-middleware
+
+## Functions
+
+- [expressMiddleware](functions/expressMiddleware.md)
