+++ resource/doc/api/business/service/express-middleware/functions/expressMiddleware.md
@@ -1,0 +1,33 @@
+[**@beecode/msh-error**](../../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../../README.md) / [business/service/express-middleware](../README.md) / expressMiddleware
+
+# Function: expressMiddleware()
+
+> **expressMiddleware**(`err`, `_req`, `res`, `_next`): `unknown`
+
+Defined in: [packages/error/src/business/service/express-middleware.ts:5](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/business/service/express-middleware.ts#L5)
+
+## Parameters
+
+### err
+
+`unknown`
+
+### \_req
+
+`Request`
+
+### res
+
+`Response`
+
+### \_next
+
+`NextFunction`
+
+## Returns
+
+`unknown`
