+++ resource/doc/api/index/README.md
@@ -1,0 +1,19 @@
+[**@beecode/msh-error**](../README.md)
+
+***
+
+[@beecode/msh-error](../README.md) / index
+
+# index
+
+## References
+
+### errorFactory
+
+Re-exports [errorFactory](../business/service/error-factory/functions/errorFactory.md)
+
+***
+
+### errorHOF
+
+Re-exports [errorHOF](../business/service/error-hof/functions/errorHOF.md)
