--- resource/doc/api/modules/error_base_model.md
@@ -1,23 +1,0 @@
-[@beecode/msh-error](../README.md) / error/base-model
-
-# Module: error/base-model
-
-## Table of contents
-
-### Classes
-
-- [ErrorBaseModel](../classes/error_base_model.ErrorBaseModel.md)
-
-### Type Aliases
-
-- [ErrorPayload](error_base_model.md#errorpayload)
-
-## Type Aliases
-
-### ErrorPayload
-
-Ƭ **ErrorPayload**: `Record`\<`string`, `unknown`\>
-
-#### Defined in
-
-[packages/error/src/error/base-model.ts:4](https://github.com/beecode-rs/msh-error/blob/744dc1e/src/error/base-model.ts#L4)
