--- resource/doc/api/modules/http_response_code.md
@@ -1,9 +1,0 @@
-[@beecode/msh-error](../README.md) / http-response-code
-
-# Module: http-response-code
-
-## Table of contents
-
-### Enumerations
-
-- [HttpResponseCodeMapper](../enums/http_response_code.HttpResponseCodeMapper.md)
