--- resource/doc/api/modules/index.md
@@ -1,22 +1,0 @@
-[@beecode/msh-error](../README.md) / index
-
-# Module: index
-
-## Table of contents
-
-### References
-
-- [errorFactory](index.md#errorfactory)
-- [errorHOF](index.md#errorhof)
-
-## References
-
-### errorFactory
-
-Re-exports [errorFactory](error_factory.md#errorfactory)
-
-___
-
-### errorHOF
-
-Re-exports [errorHOF](error_hof.md#errorhof)
