--- resource/doc/api/modules/util_string_util.md
@@ -1,25 +1,0 @@
-[@beecode/msh-error](../README.md) / util/string-util
-
-# Module: util/string-util
-
-## Table of contents
-
-### Variables
-
-- [stringUtil](util_string_util.md#stringutil)
-
-## Variables
-
-### stringUtil
-
-• `Const` **stringUtil**: `Object`
-
-#### Type declaration
-
-| Name | Type |
-| :------ | :------ |
-| `codeNameFromEnum` | (`codeEnum`: [`HttpResponseCodeMapper`](../enums/http_response_code.HttpResponseCodeMapper.md)) => `string` |
-
-#### Defined in
-
-[packages/error/src/util/string-util.ts:3](https://github.com/beecode-rs/msh-error/blob/744dc1e/src/util/string-util.ts#L3)
