+++ resource/doc/api/types/global/README.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-error**](../../README.md)
+
+***
+
+[@beecode/msh-error](../../README.md) / types/global
+
+# types/global
+
+## Variables
+
+- [esmImportMocked](variables/esmImportMocked.md)
