+++ resource/doc/api/types/global/variables/esmImportMocked.md
@@ -1,0 +1,31 @@
+[**@beecode/msh-error**](../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../README.md) / [types/global](../README.md) / esmImportMocked
+
+# Variable: esmImportMocked
+
+> `const` **esmImportMocked**: (`metaUrl`, `modulePath`) => `any`
+
+Defined in: [packages/error/src/types/global.d.ts:8](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/types/global.d.ts#L8)
+
+Example usage: esmImportMocked(import.meta.url, '#src/some-service/some-esm-module')
+
+## Parameters
+
+### metaUrl
+
+`string` \| `URL`
+
+first parameter needs to be `import.meta.url`
+
+### modulePath
+
+`string`
+
+module path absolute or relative
+
+## Returns
+
+`any`
