+++ resource/doc/api/util/http-response-code-util/README.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-error**](../../README.md)
+
+***
+
+[@beecode/msh-error](../../README.md) / util/http-response-code-util
+
+# util/http-response-code-util
+
+## Variables
+
+- [httpResponseCodeUtil](variables/httpResponseCodeUtil.md)
