+++ resource/doc/api/util/http-response-code-util/variables/httpResponseCodeUtil.md
@@ -1,0 +1,27 @@
+[**@beecode/msh-error**](../../../README.md)
+
+***
+
+[@beecode/msh-error](../../../README.md) / [util/http-response-code-util](../README.md) / httpResponseCodeUtil
+
+# Variable: httpResponseCodeUtil
+
+> `const` **httpResponseCodeUtil**: `object`
+
+Defined in: [packages/error/src/util/http-response-code-util.ts:3](https://github.com/beecode-rs/msh-error/blob/7eb9f5b703d31d76d8e758ab1268ceaf61edc1b5/src/util/http-response-code-util.ts#L3)
+
+## Type Declaration
+
+### codeNameFromEnum
+
+> **codeNameFromEnum**: (`codeEnum`) => `string`
+
+#### Parameters
+
+##### codeEnum
+
+[`HttpResponseCodeMapper`](../../../business/model/http-response-code-mapper/enumerations/HttpResponseCodeMapper.md)
+
+#### Returns
+
+`string`
