--- tsconfig.build.cjs.json
@@ -1,9 +1,0 @@
-{
-	"extends": "./tsconfig.build.json",
-	"compilerOptions": {
-		"ignoreDeprecations": "6.0",
-		"module": "CommonJS",
-		"moduleResolution": "bundler",
-		"outDir": "./lib"
-	}
-}
